LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/mo-minsize-maxsize-001-expected.txt


PASS minsize < 0 is treated as 0
PASS maxsize < minsize is treated as maxsize = minsize
PASS minsize < maxsize < 0 is treated as maxsize = minsize = 0
FAIL target size = 0 is treated as Tascent = Tdescent = minsize/2 assert_approx_equals: expected 100 +/- 1 but got 50
⥯

⥯

⥯

⥯
